backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==4.2.11
django-cors-headers==4.3.1
django-sslserver==0.22
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
pillow==10.3.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyotp==2.9.0
python-dotenv==1.0.1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
twisted==24.3.0
txaio==23.1.1
typing-extensions==4.11.0
urllib3==2.2.1
zope.interface==6.4.post2
# psycopg2==2.9.9
psycopg2-binary==2.9.9
# gunicorn==22.0.0
# uvicorn==0.30.3